Use test cases for Svelte <Navigating>
