+++ packages/svelte/tests/cases/navigating/mount/app.html
@@ -1,0 +1,13 @@
+<Navigating component={ Cancel } />
+
+<script>
+  import Navigating from "../../../../src/Navigating.html";
+  import Cancel from "../../../fixtures/Cancel.html";
+
+  export default {
+    components: { Navigating },
+    data() {
+      return { Cancel };
+    }
+  }  
+</script>
